Change to int64

+++ source/common/common/posix/thread_impl.h
@@ -14,7 +14,7 @@ public:
   ThreadIdImplPosix(int64_t id);
 
   // Thread::ThreadId
-  int tid() const override { return id_; };
+  int64 tid() const override { return id_; };
   std::string debugString() const override;
   bool isCurrentThreadId() const override;
