+++ source/common/common/win32/thread_impl.h
@@ -18,7 +18,7 @@ public:
   ThreadIdImplWin32(DWORD id);
 
   // Thread::ThreadId
-  int tid() const override { return id_; };
+  int64 tid() const override { return id_; };
   std::string debugString() const override;
   bool isCurrentThreadId() const override;
